requirements.txt
cryptography==44.0.1
numpy==1.26.4
sounddevice>=0.4.4
pygame==2.6.1
PyQt5==5.15.11
opencv-python-headless==4.11.0.86
soxr==0.5.0.post1
psutil==7.0.0
pillow==11.3.0
webrtcvad-wheels==2.0.14
sherpa-onnx>=1.10.0
pendulum==3.1.0
openai==1.86.0
opuslib==3.0.1
paho-mqtt==2.1.0
pynput==1.8.1
pyperclip==1.9.0
pypinyin==0.55.0
aiohttp==3.12.13
websockets==11.0.3
colorlog==6.9.0
qasync==0.27.1
py-machineid==0.8.0
mutagen==1.47.0
beautifulsoup4==4.12.3
brotli==1.1.0
python-dateutil==2.9.0.post0
lunar_python==1.4.4
pyinstaller==6.15.0
rich==14.1.0
packaging==25.0
requests==2.32.3